feature: add interfaces
