+++ src/main/java/HistoryManager.java
@@ -1,0 +1,6 @@
+public interface HistoryManager {
+    //Помечать задачи, как просмотренные
+    void addTask(Task task);
+    //Возвращать список просмотренных задач
+    void getHistory();
+}
