+++ src/main/java/InMemoryHistoryManager.java
@@ -1,0 +1,12 @@
+public class InMemoryHistoryManager implements  HistoryManager{
+    //Помечать задачи, как просмотренные
+    @Override
+    public void addTask(Task task) {
+
+    }
+    //Возвращать список просмотренных задач
+    @Override
+    public void getHistory() {
+
+    }
+}
